+++ lib/gretel/crumb.rb
@@ -12,6 +12,9 @@ module Gretel
 
     # Sets link of the breadcrumb.
     def link(text, url = nil)
+      # Transform objects to real paths.
+      url = url_for(url) if url
+      
       links << Gretel::Link.new(key, text, url)
     end